feat: Перенесены обработчики ошибок связанные с данными в классы storage.
